duplicate month date moved in to util
