+++ server/utils/dateUtils.ts
@@ -4,6 +4,25 @@ import ClockTime from './clockTime'
 
 type casing = 'lowercase' | 'capitalized'
 export default class DateUtils {
+  static monthNameToNumber(monthName: string): number | null {
+    const months: Record<string, number> = {
+      january: 0,
+      february: 1,
+      march: 2,
+      april: 3,
+      may: 4,
+      june: 5,
+      july: 6,
+      august: 7,
+      september: 8,
+      october: 9,
+      november: 10,
+      december: 11,
+    }
+
+    return months[monthName.toLowerCase()] ?? null
+  }
+
   static formattedDayOfWeek(inputDate: Date | string): string {
     let date: Date
     if (inputDate instanceof Date) {
